+++ lib/http/request.rb
@@ -26,9 +26,6 @@ module HTTP
     # The scheme of given URI was not understood
     class UnsupportedSchemeError < RequestError; end
 
-    # The URI given was not valid
-    class InvalidURIError < RequestError; end
-
     # Default User-Agent header value
     USER_AGENT = "http.rb/#{HTTP::VERSION}".freeze
 
@@ -335,10 +332,11 @@ module HTTP
     # @return [void]
     # @api private
     def parse_uri!(uri)
+      raise ArgumentError, "uri is nil" if uri.nil?
+      raise ArgumentError, "uri is empty" if uri.is_a?(String) && uri.empty?
+
       @uri = @uri_normalizer.call(uri)
       @scheme = @uri.scheme.to_s.downcase.to_sym if @uri.scheme
-    rescue TypeError, Addressable::URI::InvalidURIError
-      raise InvalidURIError, "invalid URI: #{uri.inspect}"
     end
 
     # Validate HTTP method and URI scheme
@@ -346,7 +344,7 @@ module HTTP
     # @api private
     def validate_method_and_scheme!
       raise(UnsupportedMethodError, "unknown method: #{verb}") unless METHODS.include?(@verb)
-      raise(InvalidURIError, "invalid URI: #{@uri}") unless @scheme
+      raise(HTTP::URI::InvalidError, "invalid URI: #{@uri}") unless @scheme
       raise(UnsupportedSchemeError, "unknown scheme: #{scheme}") unless SCHEMES.include?(@scheme)
     end